Refactored with a reusable CollapseSection component in the App in place of individual collapse components. #11
