+++ client/containers/Chart.js
@@ -1,11 +1,8 @@
-import React, { Component, Fragment } from 'react'
-import { Row, Collapse } from 'reactstrap'
+import React, { Component } from 'react'
+import { Row } from 'reactstrap'
 import Chart from 'chart.js'
 
 const styles = {
-  header: {
-    padding: '0 50px'
-  },
   row: {
     minHeight: '500px',
     padding: '0 50px'
@@ -13,23 +10,12 @@ const styles = {
   chart: {
     height: '450px',
     width: '1050px'
-  },
-  icon: {
-    color: 'rgb(133, 133, 133)',
-    cursor: 'pointer',
-    verticalAlign: 'middle',
-    lineHeight: '24px'
   }
 }
 
 export default class DataChart extends Component {
   constructor(props) {
     super(props)
-    this.state = { collapse: false }
-  }
-
-  toggle = () => {
-    this.setState({ collapse: !this.state.collapse })
   }
 
   componentDidMount() {
@@ -126,22 +112,10 @@ export default class DataChart extends Component {
   }
 
   render() { 
-    const icon = this.state.collapse
-    ? 'far fa-plus-square'
-    : 'far fa-minus-square'
-
     return (
-      <Fragment>
-        <Row style={styles.header}>
-          <h5 className="mb-3 mr-2">Chart</h5>
-          <i className={icon} style={styles.icon} onClick={this.toggle}></i>
-        </Row>
-        <Collapse isOpen={!this.state.collapse}>
-          <Row style={styles.row}>
-            <canvas ref="canvas" style={styles.chart}></canvas>
-          </Row>
-        </Collapse>
-      </Fragment>
+      <Row style={styles.row}>
+        <canvas ref="canvas" style={styles.chart}></canvas>
+      </Row>
     )
   }
 }
